+++ ddprof-lib/src/main/cpp/hotspot/vmStructs.cpp
@@ -7,7 +7,7 @@
 #include <string.h>
 #include <unistd.h>
 #include <stdarg.h>
-#include "hotspot/vmStructs.h"
+#include "hotspot/vmStructs.inline.h"
 #include "vmEntry.h"
 #include "jniHelper.h"
 #include "jvmHeap.h"
@@ -79,7 +79,7 @@ DECLARE_LONG_CONSTANTS_DO(INIT_LONG_CONSTANT, INIT_LONG_CONSTANT)
 #undef INIT_INT_CONSTANT
 #undef INIT_LONG_CONSTANT
 
-
+jfieldID VMStructs::_eetop = NULL;
 jfieldID VMStructs::_klass = NULL;
 intptr_t VMStructs::_env_offset = -1;
 void* VMStructs::_java_thread_vtbl[6];
@@ -605,26 +605,27 @@ void VMStructs::checkNativeBinding(jvmtiEnv *jvmti, JNIEnv *jni,
     jvmti->Deallocate((unsigned char *)method_name);
 }
 
-void VMThread::initialize(void* current) {
-    VMThread* vm_thread = VMThread::cast(current);
+void* VMThread::initialize(jthread thread) {
+    JNIEnv* env = VM::jni();
+    jclass thread_class = env->FindClass("java/lang/Thread");
+    if (thread_class == NULL) {
+        env->ExceptionClear();
+        return nullptr;
+    }
+
+    // Get eetop field - a bridge from Java Thread to VMThread
+    if ((_eetop = env->GetFieldID(thread_class, "eetop", "J")) == NULL) {
+        // No such field - probably not a HotSpot JVM
+        env->ExceptionClear();
+        return nullptr;
+    }
+
+    VMThread* vm_thread = fromJavaThread(env, thread);
     assert(vm_thread != nullptr);
     _has_native_thread_id = _thread_osthread_offset >= 0 && _osthread_id_offset >= 0;
-    JNIEnv* env = VM::jni();
     _env_offset = (intptr_t)env - (intptr_t)vm_thread;
     memcpy(_java_thread_vtbl, vm_thread->vtable(), sizeof(_java_thread_vtbl));
-}
-
-VMThread* VMThread::current() {
-    return VMThread::cast(JVMThread::current());
-}
-
-int VMThread::nativeThreadId(JNIEnv* jni, jthread thread) {
-//    assert(_has_native_thread_id);
-    if (_has_native_thread_id) {
-        VMThread* vm_thread = cast(JVMThread::fromJavaThread(jni, thread));
-        return vm_thread != NULL ? vm_thread->osThreadId() : -1;
-    }
-    return -1;
+    return (void*)vm_thread;
 }
 
 int VMThread::osThreadId() {
